+++ crates/sui/src/client_ptb/ptb_builder/parse_ptb.rs
@@ -662,8 +662,7 @@ impl<'a> ValueParser<'a> {
 
 #[cfg(test)]
 mod tests {
-    use crate::ptb::ptb_builder::context::FileScope;
-    use crate::ptb::ptb_builder::parse_ptb::PTBParser;
+    use crate::client_ptb::ptb_builder::{context::FileScope, parse_ptb::PTBParser};
 
     fn dummy_file_scope() -> FileScope {
         FileScope {
